+++ app/layout.tsx
@@ -3,7 +3,6 @@ import "src/assets/css/index.css";
 import "src/assets/fonts/Alfreda/stylesheet.css";
 import "src/assets/fonts/Belwe/stylesheet.css";
 import "src/assets/fonts/GTWalsheimPro/stylesheet.css";
-import "src/datadog";
 import { GoogleTagManager } from "@next/third-parties/google";
 import type { Metadata } from "next";
 import config from "src/config";
