+++ components/layout/flex/flex.type.ts
@@ -1,18 +1,4 @@
-export type Direction = "row" | "row-reverse" | "col" | "col-reverse";
+import { type VariantProps } from "tailwind-variants";
+import { flexVariants } from "./flex";
 
-export type Wrap = "wrap" | "wrap-reverse" | "nowrap";
-
-export type JustifyContent = "normal" | "start" | "end" | "center" | "between" | "around" | "evenly" | "stretch";
-
-export type AlignContent =
-  | "normal"
-  | "center"
-  | "start"
-  | "end"
-  | "between"
-  | "around"
-  | "evenly"
-  | "baseline"
-  | "stretch";
-
-export type AlignItems = "start" | "end" | "center" | "baseline" | "stretch";
+export type FlexVariants = VariantProps<typeof flexVariants>;
