+++ components/layout/icon/custom-icon-names.type.ts
@@ -1,0 +1,1 @@
+export type CustomIconsName = "dollar";
