+++ components/layout/typography/typography.type.ts
@@ -1,13 +1,4 @@
-export type TypographyVariant =
-  | "title-xl"
-  | "title-l"
-  | "title-m"
-  | "title-s"
-  | "body-l"
-  | "body-l-bold"
-  | "body-m"
-  | "body-m-bold"
-  | "body-s"
-  | "body-s-bold"
-  | "body-xs"
-  | "body-xs-bold";
+import { typographyVariants } from "./typography";
+import { type VariantProps } from "tailwind-variants";
+
+export type TypographyVariants = VariantProps<typeof typographyVariants>;
